+++ tutorial/driver_app/Example_Driver.F90
@@ -11,8 +11,7 @@ program Example_Driver
    type (MAPL_CapOptions) :: cap_options
    integer :: status
 
-   cap_options = FlapCLI(description = 'GEOS AGCM', &
-                             authors = 'GMAO')
+   cap_options = FargparseCLI()
    cap = MAPL_Cap('example', cap_options = cap_options)
    call cap%run(_RC)